Updated login functionality
