+++ backend/.env
@@ -1,7 +1,7 @@
 PORT=5000
 DB_HOST=localhost
 DB_USER=postgres
-DB_PASSWORD=equipe_06
-DB_NAME=codify_db
+DB_PASSWORD=mysecretpassword
+DB_NAME=codify_database
 DB_PORT=5432
 JWT_SECRET=key
