+++ backend/app.js
@@ -1,6 +1,7 @@
-const express = require('express');
-const dotenv = require('dotenv');
-const accountRoutes = require('./src/routes/accountRoutes');
+const express = require("express");
+const dotenv = require("dotenv");
+const accountRoutes = require("./src/routes/accountRoutes");
+const authRoutes = require("./src/routes/authRoutes");
 
 dotenv.config();
 
@@ -10,11 +11,12 @@ const PORT = process.env.PORT || 5000;
 app.use(express.json());
 
 // Routes
-app.use('/api', accountRoutes);
+app.use("/api", accountRoutes);
+app.use("/api", authRoutes);
 
 // Test endpoint
-app.get('/', (req, res) => {
-  res.send('Hello, this is the backend!');
+app.get("/", (req, res) => {
+  res.send("Hello, this is the backend!");
 });
 
 // Start the server
