+++ backend/knexfile.js
@@ -1,59 +1,45 @@
-require('dotenv').config(); // Load environment variables from .env
+// Update with your config settings.
 
+/**
+ * @type { Object.<string, import("knex").Knex.Config> }
+ */
 module.exports = {
   development: {
-    client: 'pg', 
+    client: "sqlite3",
     connection: {
-      host: process.env.DB_HOST, 
-      user: process.env.DB_USER,
-      password: process.env.DB_PASSWORD, 
-      database: process.env.DB_NAME, 
-      port: process.env.DB_PORT, 
-    },
-    migrations: {
-      directory: './migrations', 
-      tableName: 'knex_migrations', 
-    },
-    seeds: {
-      directory: './seeds', 
+      filename: "./dev.sqlite3",
     },
   },
 
   staging: {
-    client: 'pg',
+    client: "postgresql",
     connection: {
-      host: process.env.DB_HOST,
-      user: process.env.DB_USER,
-      password: process.env.DB_PASSWORD,
-      database: process.env.DB_NAME,
-      port: process.env.DB_PORT,
+      database: "codify_database",
+      user: "postgres",
+      password: "mysecretpassword",
     },
     pool: {
       min: 2,
       max: 10,
     },
     migrations: {
-      directory: './migrations',
-      tableName: 'knex_migrations',
+      tableName: "knex_migrations",
     },
   },
 
   production: {
-    client: 'pg',
+    client: "postgresql",
     connection: {
-      host: process.env.DB_HOST,
-      user: process.env.DB_USER,
-      password: process.env.DB_PASSWORD,
-      database: process.env.DB_NAME,
-      port: process.env.DB_PORT,
+      database: "codify_database",
+      user: "postgres",
+      password: "mysecretpassword",
     },
     pool: {
       min: 2,
       max: 10,
     },
     migrations: {
-      directory: './migrations',
-      tableName: 'knex_migrations',
+      tableName: "knex_migrations",
     },
   },
 };
