+++ backend/src/utils/tokenUtils.js
@@ -1,14 +1,14 @@
-const jwt = require('jsonwebtoken');
-require('dotenv').config(); 
+const jwt = require("jsonwebtoken");
+require("dotenv").config();
 
-const SECRET_KEY = process.env.JWT_SECRET || 'your_secret_key'; 
+const SECRET_KEY = process.env.JWT_SECRET || "your_secret_key";
 
 // Generate JWT Token
 const generateToken = (user) => {
   return jwt.sign(
     { id: user.id, email: user.email, role: user.user_role },
     SECRET_KEY,
-    { expiresIn: '2d' } // Token expires in 2 days
+    { expiresIn: "2d" } // Token expires in 2 days
   );
 };
